Title: 숫자 사이의 배수, Time: 5ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240404/숫자 사이의 배수/multiplication-between-numbers.cpp
@@ -18,7 +18,7 @@ int main() {
         }
     }
 
-    cout << sum << " " << sum/cnt;
+    cout << sum << " " << (double)sum/cnt;
 
     return 0;
 }
